+++ packages/enhanced/test/configCases/tree-shaking-share/reshake-share/CustomPlugin.js
@@ -1,0 +1,22 @@
+class CustomPlugin {
+  apply(compiler) {
+    compiler.hooks.thisCompilation.tap('applyPlugins', (compilation) => {
+      compilation.hooks.processAssets.tapPromise(
+        {
+          name: 'applyPlugins',
+        },
+        async () => {
+          compilation.emitAsset(
+            'apply-plugin.json',
+            new compilation.compiler.webpack.sources.RawSource(
+              JSON.stringify({
+                secondary: true,
+              }),
+            ),
+          );
+        },
+      );
+    });
+  }
+}
+module.exports = CustomPlugin;