+++ packages/esbuild/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @module-federation/esbuild
 
+## 0.0.92
+
+### Patch Changes
+
+- @module-federation/sdk@0.23.0
+
 ## 0.0.91
 
 ### Patch Changes