+++ packages/metro-plugin-rnc-cli/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @module-federation/metro-plugin-rnc-cli
 
+## 0.23.0
+
+### Patch Changes
+
+- @module-federation/metro@0.23.0
+
 ## 0.22.1
 
 ### Patch Changes